add: 모달 컴포넌트 추상화
