+++ next-app/src/components/common/Modal/index.ts
@@ -1,2 +1,3 @@
 export { useModal } from './useModal'
+export { DimmerLayout } from './DimmerLayout'
 export { ModalLayout } from './ModalLayout'
