+++ next-app/src/components/views/RssInput/RssInputContainer.tsx
@@ -1,17 +1,41 @@
 import Image from 'next/legacy/image'
-import { useRef } from 'react'
 
 import Flex from 'components/common/Flex'
 import Input from './Input'
 import { isRssUrlValid } from './RssInputContainer.utils'
 import { useRssInput } from './hooks'
+import { ModalLayout, useModal } from 'components/common/Modal'
+import Toast from 'components/common/Toast'
+import Button from 'components/common/Button'
 
 import * as S from './RssInputContainer.style'
 
 import Icons from 'assets/icons'
 
 const RssInputContainer = () => {
-  const inputRef = useRef<HTMLInputElement>(null)
+  const { handleOpen, renderModal } = useModal({
+    content: (
+      <ModalLayout title="RSS 수동으로 추가하기">
+        <Flex direction="column" gap={12}>
+          <input placeholder="블로그 URL을 입력해주세요" />
+          <input placeholder="RSS URL을 입력해주세요" />
+        </Flex>
+        <Flex justify="end">
+          <Button
+            buttonStyle="secondary"
+            onClick={() => {
+              Toast.show({
+                content: 'RSS 주소가 성공적으로 추가되었습니다.',
+              })
+            }}
+          >
+            RSS 추가하기
+          </Button>
+        </Flex>
+      </ModalLayout>
+    ),
+  })
+
   const { url, onSubmit, handleInput, isSubmitting } = useRssInput()
 
   const isSubmitEnabled = !isSubmitting && isRssUrlValid(url)
@@ -21,12 +45,30 @@ const RssInputContainer = () => {
       <S.Form onSubmit={(e) => isSubmitEnabled && onSubmit(e)}>
         <Flex justify="center" align="center" style={{ position: 'relative' }}>
           <Input
-            ref={inputRef}
             name="url"
             placeholder="URL을 추가해서 피드로 모아보세요!"
             isError={!!url && !isRssUrlValid(url)}
             onChange={handleInput}
             value={url}
+            renderInputIcon={({ selectedValue, clearValue }) => {
+              return selectedValue ? (
+                <Image
+                  alt="삭제 버튼"
+                  src={Icons.CancelCircle}
+                  width={24}
+                  height={24}
+                  onClick={clearValue}
+                />
+              ) : (
+                <Image
+                  alt="RSS 직접 추가"
+                  src={Icons.RssCircle}
+                  width={24}
+                  height={24}
+                  onClick={handleOpen}
+                />
+              )
+            }}
           />
           <S.AddButton
             type="submit"
@@ -44,11 +86,13 @@ const RssInputContainer = () => {
           {!!url && isRssUrlValid(url) === false && (
             <S.Error>
               RSS 추가를 할 수 없는 형식의 링크입니다.{' '}
-              <S.UnderLine>RSS 링크를 찾는 법</S.UnderLine>을 참고해주세요.
+              <S.UnderLine onClick={handleOpen}>RSS 링크를 찾는 법</S.UnderLine>
+              을 참고해주세요.
             </S.Error>
           )}
         </Flex>
       </S.Form>
+      {renderModal()}
     </S.Container>
   )
 }
