--- next-app/src/components/views/RssInput/modal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RssInputModal'
